clean up server-side code
